04_docker_compose/01_flask-postgres-redis/requirements.txt
Flask==2.0.1
redis==3.5.3
psycopg2-binary==2.9.1